Make greyVest the default theme
Remove blueberry, material, and merge base theme into grey-vest

Had to break down ./src/utils/theme.js into multiple files because of
circular imports
