+++ packages/react/package.json
@@ -29,8 +29,7 @@
   "scripts": {
     "prepack": "node ../../scripts/esbuild.js --platform=browser --loader:.js=jsx",
     "storybook": "storybook dev -p 6006",
-    "storybook:build": "storybook build",
-    "update-pkg": "node scripts/updatePackage.js"
+    "storybook:build": "storybook build"
   },
   "repository": {
     "type": "git",
@@ -55,8 +54,6 @@
   },
   "devDependencies": {
     "@date-io/moment": "^1.3.9",
-    "@material-ui/core": "^4.3.3",
-    "@material-ui/pickers": "^3.2.3",
     "@storybook/addon-essentials": "^7.0.0-rc.3",
     "@storybook/addon-links": "^7.0.0-rc.3",
     "@storybook/blocks": "^7.0.0-rc.3",
@@ -69,7 +66,6 @@
     "elasticsearch-browser": "^14.2.2",
     "emoji-datasource": "^5.0.1",
     "eslint-plugin-react": "^7.32.0",
-    "material-ui-chip-input": "^1.0.0",
     "mobx": "^4.3.1",
     "mobx-react": "^6.3.0",
     "mobx-utils": "^5.0.0",
@@ -78,9 +74,7 @@
     "react-dom": "^16.8.0",
     "react-select": "^2.0.0",
     "react-test-renderer": "^16.2.0",
-    "read-pkg": "^5.2.0",
-    "storybook": "^7.0.0-rc.3",
-    "write-pkg": "^4.0.0"
+    "storybook": "^7.0.0-rc.3"
   },
   "dependencies": {
     "futil": "^1.71.8",
