+++ packages/react/src/themes/greyVest/index.js
@@ -19,12 +19,14 @@ import {
   TextButton,
   TextHighlight,
   TextInput,
+  Popover,
 } from '../../greyVest/index.js'
 
 //components used only for theme
 import PickerItem from './PickerItem.js'
 import TagsInput, { Tag } from './TagsInput.js'
 import Root from './Root.js'
+import UnmappedNodeComponent from './UnmappedNodeComponent.js'
 
 export default {
   AlternateButton: TextButton,
@@ -48,4 +50,13 @@ export default {
   Select,
   Table,
   TextHighlight,
+  Popover,
+  Tbody: 'tbody',
+  Td: 'td',
+  TextButton: 'button',
+  Tfoot: 'tfoot',
+  Th: 'th',
+  Thead: 'thead',
+  Tr: 'tr',
+  UnmappedNodeComponent,
 }
