+++ packages/react/src/themes/index.js
@@ -1,5 +1,3 @@
-import base from './base/index.js'
-import blueberry from './blueberry/index.js'
 import greyVest from './greyVest/index.js'
 
-export { base, blueberry, greyVest }
+export { greyVest }
