--- packages/react/src/themes/material/Style.js
@@ -1,18 +1,0 @@
-import React from 'react'
-import GVStyle from '../greyVest/Style.js'
-
-export default () => (
-  <>
-    <style>
-      {`
-        .material-table th span {
-          display: flex;
-          flex-wrap: nowrap;
-          align-items: center;
-          justify-content: center;
-        }
-      `}
-    </style>
-    <GVStyle />
-  </>
-)
