--- packages/react/src/themes/material/utils.js
@@ -1,6 +1,0 @@
-import F from 'futil'
-
-export let openBinding = (...lens) => ({
-  open: F.view(...lens),
-  onClose: F.off(...lens),
-})
