+++ packages/react/src/utils/theme/context.js
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export default React.createContext()
